Adding sobel functionality
stackImages and useClassifier are now using sobel versions of the
images. Very slow at the moment, need to work out downsampling so that
the sobel filter doesn't take so long. soble.py creates a sobel image
for testing the classifier. So far error has not reduced compared with
straight RGB, and that's using RGB WITH sobel.